Upgrade to 2021
Upgraded robot code to 2021. Builds successfully theoretically, needs testing. Removed colorwheel code. Hook Extension code should be removed soon too.

+++ src/main/java/frc/robot/Wheels.java
@@ -29,12 +29,12 @@ public class Wheels {
 
     public double getRotations(String location) { 
         if (location == "fL") {
-            int temp = frontLeft.getSelectedSensorPosition();
+            double temp = frontLeft.getSelectedSensorPosition();
             double val = temp / 4096;
             return val;
         }
         if (location == "bR") {
-            int temp = backRight.getSelectedSensorPosition();
+            double temp = backRight.getSelectedSensorPosition();
             double val = temp / 4096;
             return val;
